payment provider implementation class location moved and minor update in data type specification
