+++ src/core/domain/payment/payment.types.ts
@@ -3,6 +3,10 @@ export type CreatePaymentProps = {
   amount: number;
 };
 
+export type UpdatePaymentProps =
+  | { encodedData: string; pidx?: string }
+  | { encodedData?: string; pidx: string };
+
 export type PaymentResponseProps = {
   transactionId: string;
   totalAmount: number;
