+++ src/core/ports/in/payment/payment-usecase.port.ts
@@ -1,7 +1,7 @@
 import { Payment } from '../../../domain/payment/payment.domain';
-import { PaymentResponseProps } from '../../../domain/payment/payment.types';
+import { PaymentResponseProps, UpdatePaymentProps } from '../../../domain/payment/payment.types';
 
 export abstract class PaymentUseCase {
   abstract initPayment(payment:Payment): Promise<any>;
-  abstract verify(data:any): Promise<PaymentResponseProps>;
+  abstract verify(data:UpdatePaymentProps): Promise<PaymentResponseProps>;
 }
