+++ src/frameworks/primary/dto/request/transaction/transaction.dto.ts
@@ -66,6 +66,7 @@ export class UpdateTransactionDto {
   @ApiProperty()
   @IsOptional()
   encodedData: string;
+
   @ApiProperty()
   @IsOptional()
   pidx: string;
